(Map) added grass and water assets, added positions tools and added possibility to set the top-left display corner using floats numbers to make beautiful moves (was with integer numbers)

+++ src/map/map_gen.c
@@ -14,9 +14,9 @@ map_t	*generate_map(int seed)
 	map_t *map = init_map("test map", seed);
 	chunk_t *_chunk = NULL;
 
-	for (int y = -5; y < 6; y++) {
-		for (int x = -5; x < 6; x++) {
-			_chunk = chunk_gen(map, (pos_t){x, y, 0});
+	for (float y = -5.0; y < 6.0; y++) {
+		for (float x = -5.0; x < 6.0; x++) {
+			_chunk = chunk_gen(map, (pos_t){x, y, 0.0});
 			add_to_linked_list(&map->chunks, _chunk);
 		}
 	}
